create BaseCooker
