+++ src/plugins/trace/dll/Tables/QuicApiTable.cs
@@ -6,11 +6,10 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
-using LTTngCds.CookerData;
 using Microsoft.Performance.SDK;
 using Microsoft.Performance.SDK.Extensibility;
 using Microsoft.Performance.SDK.Processing;
+using QuicTrace.Cookers;
 using QuicTrace.DataModel;
 
 namespace QuicTrace.Tables
@@ -23,7 +22,7 @@ namespace QuicTrace.Tables
            "QUIC API Calls",
            "QUIC API Calls",
            category: "System Activity",
-           requiredDataCookers: new List<DataCookerPath> { QuicEventCooker.CookerPath });
+           requiredDataCookers: new List<DataCookerPath> { QuicEtwEventCooker.CookerPath });
 
         private static readonly ColumnConfiguration typeColumnConfig =
             new ColumnConfiguration(
@@ -115,7 +114,7 @@ namespace QuicTrace.Tables
         public static bool IsDataAvailable(IDataExtensionRetrieval tableData)
         {
             Debug.Assert(!(tableData is null));
-            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEventCooker.CookerPath, "State"));
+            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEtwEventCooker.CookerPath, "State"));
             return quicState != null && quicState.DataAvailableFlags.HasFlag(QuicDataAvailableFlags.Api);
         }
 
@@ -123,7 +122,7 @@ namespace QuicTrace.Tables
         {
             Debug.Assert(!(tableBuilder is null) && !(tableData is null));
 
-            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEventCooker.CookerPath, "State"));
+            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEtwEventCooker.CookerPath, "State"));
             if (quicState == null)
             {
                 return;
