+++ src/plugins/trace/dll/Tables/QuicTxBlockedTable.cs
@@ -10,6 +10,7 @@ using System.Linq;
 using Microsoft.Performance.SDK;
 using Microsoft.Performance.SDK.Extensibility;
 using Microsoft.Performance.SDK.Processing;
+using QuicTrace.Cookers;
 using QuicTrace.DataModel;
 
 namespace QuicTrace.Tables
@@ -22,7 +23,7 @@ namespace QuicTrace.Tables
            "QUIC TX Blocked State",
            "QUIC TX Blocked State",
            category: "Communications",
-           requiredDataCookers: new List<DataCookerPath> { QuicEventCooker.CookerPath });
+           requiredDataCookers: new List<DataCookerPath> { QuicEtwEventCooker.CookerPath });
 
         private static readonly ColumnConfiguration connectionColumnConfig =
             new ColumnConfiguration(
@@ -115,7 +116,7 @@ namespace QuicTrace.Tables
         public static bool IsDataAvailable(IDataExtensionRetrieval tableData)
         {
             Debug.Assert(!(tableData is null));
-            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEventCooker.CookerPath, "State"));
+            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEtwEventCooker.CookerPath, "State"));
             return quicState != null && quicState.DataAvailableFlags.HasFlag(QuicDataAvailableFlags.ConnectionFlowBlocked);
         }
 
@@ -137,7 +138,7 @@ namespace QuicTrace.Tables
         {
             Debug.Assert(!(tableBuilder is null) && !(tableData is null));
 
-            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEventCooker.CookerPath, "State"));
+            var quicState = tableData.QueryOutput<QuicState>(new DataOutputPath(QuicEtwEventCooker.CookerPath, "State"));
             if (quicState == null)
             {
                 return;
